benchmarks refactor
