+++ benchmarks/src/main/scala/io/scalaland/chimney/benchmarks/ErrorAccTransformersBenchmark.scala
@@ -12,52 +12,16 @@ class ErrorAccTransformersBenchmark extends CommonBenchmarkSettings {
 
   var simple: Simple = samples.simpleSample
 
-  private final val transformerF: TransformerF[M, Simple, SimpleOutput] =
-    TransformerF
-      .define[M, Simple, SimpleOutput]
-      .withFieldComputedF(_.a, s => happy.validateA(s.a).left.map(_.map(TransformationError(_))))
-      .withFieldComputedF(_.b, s => happy.validateB(s.b).left.map(_.map(TransformationError(_))))
-      .withFieldComputedF(_.c, s => happy.validateC(s.c).left.map(_.map(TransformationError(_))))
-      .withFieldComputedF(_.d, s => happy.validateD(s.d).left.map(_.map(TransformationError(_))))
-      .buildTransformer
-
-  private final val transformerPartial: PartialTransformer[Simple, SimpleOutput] =
-    PartialTransformer
-      .define[Simple, SimpleOutput]
-      .withFieldComputedPartial(_.a, s => happy.validateA(s.a).toPartialTransformerResult)
-      .withFieldComputedPartial(_.b, s => happy.validateB(s.b).toPartialTransformerResult)
-      .withFieldComputedPartial(_.c, s => happy.validateC(s.c).toPartialTransformerResult)
-      .withFieldComputedPartial(_.d, s => happy.validateD(s.d).toPartialTransformerResult)
-      .buildTransformer
-
-  private final val transformerFUnhappy: TransformerF[M, Simple, SimpleOutput] =
-    TransformerF
-      .define[M, Simple, SimpleOutput]
-      .withFieldComputedF(_.a, s => unhappy.validateA(s.a).left.map(_.map(TransformationError(_))))
-      .withFieldComputedF(_.b, s => unhappy.validateB(s.b).left.map(_.map(TransformationError(_))))
-      .withFieldComputedF(_.c, s => unhappy.validateC(s.c).left.map(_.map(TransformationError(_))))
-      .withFieldComputedF(_.d, s => unhappy.validateD(s.d).left.map(_.map(TransformationError(_))))
-      .buildTransformer
-
-  private final val transformerPartialUnhappy: PartialTransformer[Simple, SimpleOutput] =
-    PartialTransformer
-      .define[Simple, SimpleOutput]
-      .withFieldComputedPartial(_.a, s => unhappy.validateA(s.a).toPartialTransformerResult)
-      .withFieldComputedPartial(_.b, s => unhappy.validateB(s.b).toPartialTransformerResult)
-      .withFieldComputedPartial(_.c, s => unhappy.validateC(s.c).toPartialTransformerResult)
-      .withFieldComputedPartial(_.d, s => unhappy.validateD(s.d).toPartialTransformerResult)
-      .buildTransformer
+  @Benchmark
+  def simpleHappyLiftedTransformer: M[SimpleOutput] =
+    simpleTransformerLifted.transform(simple)
 
   @Benchmark
-  def happyLiftedTransformer: M[SimpleOutput] =
-    transformerF.transform(simple)
+  def simpleHappyPartialTransformer: PartialTransformer.Result[SimpleOutput] =
+    simpleTransformerPartial.transform(simple)
 
   @Benchmark
-  def happyPartialTransformer: PartialTransformer.Result[SimpleOutput] =
-    transformerPartial.transform(simple)
-
-  @Benchmark
-  def happyLiftedTransformerInlineDsl: M[SimpleOutput] =
+  def simpleHappyInlineDslLiftedTransformer: M[SimpleOutput] =
     simple
       .intoF[M, SimpleOutput]
       .withFieldComputedF(_.a, s => happy.validateA(s.a).left.map(_.map(TransformationError(_))))
@@ -67,7 +31,7 @@ class ErrorAccTransformersBenchmark extends CommonBenchmarkSettings {
       .transform
 
   @Benchmark
-  def happyPartialTransformerInlineDsl: PartialTransformer.Result[SimpleOutput] =
+  def simpleHappyInlineDslPartialTransformer: PartialTransformer.Result[SimpleOutput] =
     simple
       .intoPartial[SimpleOutput]
       .withFieldComputedPartial(_.a, s => happy.validateA(s.a).toPartialTransformerResult)
@@ -77,8 +41,8 @@ class ErrorAccTransformersBenchmark extends CommonBenchmarkSettings {
       .transform
 
   @Benchmark
-  def happyByHandErrorAccEitherSwap: M[SimpleOutput] =
-    byHandErrorAccEitherSwap(
+  def simpleHappyByHandErrorAccEitherSwap: M[SimpleOutput] =
+    simpleByHandErrorAccEitherSwap(
       simple,
       happy.validateA(_).left.map(_.map(TransformationError(_))),
       happy.validateB(_).left.map(_.map(TransformationError(_))),
@@ -87,8 +51,8 @@ class ErrorAccTransformersBenchmark extends CommonBenchmarkSettings {
     )
 
   @Benchmark
-  def happyByHandErrorAccCrazyNesting: M[SimpleOutput] =
-    byHandErrorAccCrazyNesting(
+  def simpleHappyByHandErrorAccCrazyNesting: M[SimpleOutput] =
+    simpleByHandErrorAccCrazyNesting(
       simple,
       happy.validateA(_).left.map(_.map(TransformationError(_))),
       happy.validateB(_).left.map(_.map(TransformationError(_))),
@@ -97,15 +61,15 @@ class ErrorAccTransformersBenchmark extends CommonBenchmarkSettings {
     )
 
   @Benchmark
-  def unhappyLiftedTransformer: M[SimpleOutput] =
-    transformerFUnhappy.transform(simple)
+  def simpleUnhappyLiftedTransformer: M[SimpleOutput] =
+    simpleTransformerLiftedUnhappy.transform(simple)
 
   @Benchmark
-  def unhappyPartialTransformer: PartialTransformer.Result[SimpleOutput] =
-    transformerPartialUnhappy.transform(simple)
+  def simpleUnhappyPartialTransformer: PartialTransformer.Result[SimpleOutput] =
+    simpleTransformerPartialUnhappy.transform(simple)
 
   @Benchmark
-  def unhappyLiftedTransformerInlineDsl: M[SimpleOutput] =
+  def simpleUnhappyInlineDslLiftedTransformer: M[SimpleOutput] =
     simple
       .intoF[M, SimpleOutput]
       .withFieldComputedF(_.a, s => unhappy.validateA(s.a).left.map(_.map(TransformationError(_))))
@@ -115,7 +79,7 @@ class ErrorAccTransformersBenchmark extends CommonBenchmarkSettings {
       .transform
 
   @Benchmark
-  def unhappyPartialTransformerInlineDsl: PartialTransformer.Result[SimpleOutput] =
+  def simpleUnhappyInlineDslPartialTransformer: PartialTransformer.Result[SimpleOutput] =
     simple
       .intoPartial[SimpleOutput]
       .withFieldComputedPartial(_.a, s => unhappy.validateA(s.a).toPartialTransformerResult)
@@ -125,8 +89,8 @@ class ErrorAccTransformersBenchmark extends CommonBenchmarkSettings {
       .transform
 
   @Benchmark
-  def unhappyByHandErrorAccEitherSwap: M[SimpleOutput] =
-    byHandErrorAccEitherSwap(
+  def simpleUnhappyByHandErrorAccEitherSwap: M[SimpleOutput] =
+    simpleByHandErrorAccEitherSwap(
       simple,
       unhappy.validateA(_).left.map(_.map(TransformationError(_))),
       unhappy.validateB(_).left.map(_.map(TransformationError(_))),
@@ -135,8 +99,8 @@ class ErrorAccTransformersBenchmark extends CommonBenchmarkSettings {
     )
 
   @Benchmark
-  def unhappyByHandErrorAccCrazyNesting: M[SimpleOutput] =
-    byHandErrorAccCrazyNesting(
+  def simpleUnhappyByHandErrorAccCrazyNesting: M[SimpleOutput] =
+    simpleByHandErrorAccCrazyNesting(
       simple,
       unhappy.validateA(_).left.map(_.map(TransformationError(_))),
       unhappy.validateB(_).left.map(_.map(TransformationError(_))),
@@ -144,97 +108,40 @@ class ErrorAccTransformersBenchmark extends CommonBenchmarkSettings {
       unhappy.validateD(_).left.map(_.map(TransformationError(_)))
     )
 
-  private def byHandErrorAccEitherSwap(
-      simple: Simple,
-      fa: Int => M[Int],
-      fb: Double => M[Double],
-      fc: String => M[String],
-      fd: Option[String] => M[Option[String]]
-  ): M[SimpleOutput] = {
-    val valA = fa(simple.a)
-    val valB = fb(simple.b)
-    val valC = fc(simple.c)
-    val valD = fd(simple.d)
+  private final val simpleTransformerLifted: TransformerF[M, Simple, SimpleOutput] =
+    TransformerF
+      .define[M, Simple, SimpleOutput]
+      .withFieldComputedF(_.a, s => happy.validateA(s.a).left.map(_.map(TransformationError(_))))
+      .withFieldComputedF(_.b, s => happy.validateB(s.b).left.map(_.map(TransformationError(_))))
+      .withFieldComputedF(_.c, s => happy.validateC(s.c).left.map(_.map(TransformationError(_))))
+      .withFieldComputedF(_.d, s => happy.validateD(s.d).left.map(_.map(TransformationError(_))))
+      .buildTransformer
 
-    if(valA.isRight && valB.isRight && valC.isRight && valD.isRight) {
-      Right(SimpleOutput(valA.toOption.get, valB.toOption.get, valC.toOption.get, valD.toOption.get))
-    } else {
-      val errsB = Vector.newBuilder[TransformationError[String]]
-      errsB ++= valA.swap.getOrElse(Vector.empty)
-      errsB ++= valB.swap.getOrElse(Vector.empty)
-      errsB ++= valC.swap.getOrElse(Vector.empty)
-      errsB ++= valD.swap.getOrElse(Vector.empty)
-      Left(errsB.result())
-    }
-  }
+  private final val simpleTransformerPartial: PartialTransformer[Simple, SimpleOutput] =
+    PartialTransformer
+      .define[Simple, SimpleOutput]
+      .withFieldComputedPartial(_.a, s => happy.validateA(s.a).toPartialTransformerResult)
+      .withFieldComputedPartial(_.b, s => happy.validateB(s.b).toPartialTransformerResult)
+      .withFieldComputedPartial(_.c, s => happy.validateC(s.c).toPartialTransformerResult)
+      .withFieldComputedPartial(_.d, s => happy.validateD(s.d).toPartialTransformerResult)
+      .buildTransformer
 
-  private def byHandErrorAccCrazyNesting(
-      simple: Simple,
-      fa: Int => M[Int],
-      fb: Double => M[Double],
-      fc: String => M[String],
-      fd: Option[String] => M[Option[String]]
-  ): M[SimpleOutput] = {
-    fa(simple.a) match {
-      case Right(a) =>
-        fb(simple.b) match {
-          case Right(b) =>
-            fc(simple.c) match {
-              case Right(c) =>
-                fd(simple.d) match {
-                  case Right(d)         => Right(SimpleOutput(a, b, c, d))
-                  case retVal @ Left(_) => retVal.asInstanceOf[M[SimpleOutput]]
-                }
-              case Left(errs3) =>
-                fd(simple.d) match {
-                  case Right(_)    => Left(errs3)
-                  case Left(errs4) => Left(errs3 ++ errs4)
-                }
-            }
-          case Left(errs2) =>
-            fc(simple.c) match {
-              case Right(_) =>
-                fd(simple.d) match {
-                  case Right(_)    => Left(errs2)
-                  case Left(errs4) => Left(errs2 ++ errs4)
-                }
-              case Left(errs3) =>
-                fd(simple.d) match {
-                  case Right(_)    => Left(errs2 ++ errs3)
-                  case Left(errs4) => Left(errs2 ++ errs3 ++ errs4)
-                }
-            }
-        }
-      case Left(errs1) =>
-        fb(simple.b) match {
-          case Right(_) =>
-            fc(simple.c) match {
-              case Right(_) =>
-                fd(simple.d) match {
-                  case Right(_)    => Left(errs1)
-                  case Left(errs4) => Left(errs1 ++ errs4)
-                }
-              case Left(errs3) =>
-                fd(simple.d) match {
-                  case Right(_)    => Left(errs3)
-                  case Left(errs4) => Left(errs1 ++ errs3 ++ errs4)
-                }
-            }
-          case Left(errs2) =>
-            fc(simple.c) match {
-              case Right(_) =>
-                fd(simple.d) match {
-                  case Right(_)    => Left(errs1 ++ errs2)
-                  case Left(errs4) => Left(errs1 ++ errs2 ++ errs4)
-                }
-              case Left(errs3) =>
-                fd(simple.d) match {
-                  case Right(_)    => Left(errs1 ++ errs2 ++ errs3)
-                  case Left(errs4) => Left(errs1 ++ errs2 ++ errs3 ++ errs4)
-                }
-            }
-        }
-    }
-  }
+  private final val simpleTransformerLiftedUnhappy: TransformerF[M, Simple, SimpleOutput] =
+    TransformerF
+      .define[M, Simple, SimpleOutput]
+      .withFieldComputedF(_.a, s => unhappy.validateA(s.a).left.map(_.map(TransformationError(_))))
+      .withFieldComputedF(_.b, s => unhappy.validateB(s.b).left.map(_.map(TransformationError(_))))
+      .withFieldComputedF(_.c, s => unhappy.validateC(s.c).left.map(_.map(TransformationError(_))))
+      .withFieldComputedF(_.d, s => unhappy.validateD(s.d).left.map(_.map(TransformationError(_))))
+      .buildTransformer
+
+  private final val simpleTransformerPartialUnhappy: PartialTransformer[Simple, SimpleOutput] =
+    PartialTransformer
+      .define[Simple, SimpleOutput]
+      .withFieldComputedPartial(_.a, s => unhappy.validateA(s.a).toPartialTransformerResult)
+      .withFieldComputedPartial(_.b, s => unhappy.validateB(s.b).toPartialTransformerResult)
+      .withFieldComputedPartial(_.c, s => unhappy.validateC(s.c).toPartialTransformerResult)
+      .withFieldComputedPartial(_.d, s => unhappy.validateD(s.d).toPartialTransformerResult)
+      .buildTransformer
 
 }
